pyproject.toml
[project]
name = "data-engineering-iu-task-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "edd2abt", email = "Dennis.Eder2@de.bosch.com" }]
requires-python = ">=3.10"
dependencies = [
    "kafka-python>=2.2.15",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pymongo>=4.13.2",
    "pyyaml>=6.0.2",
]

[project.scripts]
data-engineering-iu-task-2 = "data_engineering_iu_task_2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
